// RMG/ClientApp/src/app/models/Applicant.ts
export class Applicant {

  applicant_code: string;
  applicant_name: string;
  
  gender: string;
  date_of_birth: string;
  marital_status: string;
  educational_details: string;
  experience: string;
  previous_company: string;
  employment_type: string;
  skills: string;
  phone: string;
  email_id: string;
  address: string;
  uid_type: string;
  uid_number: string;
  notice_period:string;
  certification: string;
  languages_known: string;
  applicant_status: string;



}